Fix property page server action placement issues
